simplify API logic
